# backend/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a backup backend server.",
        "title": "RockBackup API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/backup/job": {
            "post": {
                "description": "start backup job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BackupJob"
                ],
                "summary": "Start backup job",
                "parameters": [
                    {
                        "description": "backup job params",
                        "name": "policy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.StartBackupJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/service/file/get": {
            "get": {
                "description": "get backup policy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Policy"
                ],
                "summary": "Get backup policy",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PolicyView"
                        }
                    }
                }
            }
        },
        "/service/file/open": {
            "post": {
                "description": "create backup policy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Policy"
                ],
                "summary": "Create backup policy",
                "parameters": [
                    {
                        "description": "backup policy params",
                        "name": "policy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.PolicyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.PolicyRequest": {
            "type": "object",
            "required": [
                "album_name",
                "backup_cycle",
                "backup_source_id",
                "backup_source_path",
                "full_backup_schedule",
                "hostname",
                "incr_backup_schedule",
                "repository_id",
                "schedule_desc",
                "start_time"
            ],
            "properties": {
                "album_name": {
                    "type": "integer"
                },
                "backup_cycle": {
                    "type": "integer"
                },
                "backup_source_id": {
                    "type": "integer"
                },
                "backup_source_path": {
                    "type": "string"
                },
                "full_backup_schedule": {
                    "type": "string"
                },
                "hostname": {
                    "type": "string"
                },
                "incr_backup_schedule": {
                    "type": "string"
                },
                "repository_id": {
                    "type": "integer"
                },
                "schedule_desc": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "api.PolicyView": {
            "type": "object",
            "properties": {
                "backup_source_id": {
                    "type": "integer"
                },
                "hostname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "repository_id": {
                    "type": "integer"
                },
                "retention": {
                    "type": "integer"
                },
                "source_host": {
                    "type": "string"
                },
                "source_id": {
                    "type": "integer"
                },
                "source_path": {
                    "type": "string"
                },
                "source_type": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "api.Response": {
            "type": "object",
            "properties": {
                "result": {}
            }
        },
        "api.StartBackupJobRequest": {
            "type": "object",
            "properties": {
                "backup_type": {
                    "type": "string"
                },
                "policy_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}